# Conflicts:
#	TEAM PROJECT/core/src/net/uridium/game/screen/GameScreen.java
#	TEAM PROJECT/core/src/net/uridium/game/screen/LobbyScreen.java
#	TEAM PROJECT/core/src/net/uridium/game/server/LobbyServer.java

+++ TEAM PROJECT/core/src/net/uridium/game/server/msg/PlayerHealthData.java
@@ -4,10 +4,10 @@ import java.io.Serializable;
 
 public class PlayerHealthData implements Serializable {
     public int playerID;
-    public int health;
-    public int maxHealth;
+    public float health;
+    public float maxHealth;
 
-    public PlayerHealthData(int playerID, int health, int maxHealth) {
+    public PlayerHealthData(int playerID, float health, float maxHealth) {
         this.playerID = playerID;
         this.health = health;
         this.maxHealth = maxHealth;
